TMC13 -> Combine communication functionalities in prevac_protocol
